requirements.txt
Django==5.2.8
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.2
celery==5.4.0
redis==5.0.4
opencv-python==4.10.0.84
numpy==2.1.3
Pillow==10.4.0
psycopg2-binary==2.9.9
requests==2.31.0
jira==3.8.0
selenium==4.15.2
pytesseract==0.3.10
scikit-image==0.22.0
scikit-learn==1.5.2
joblib==1.4.2

